Like home,my profile, bugs fixed, Search bar impl

+++ src/main/java/com/example/demo/services/LikeServiceImpl.java
@@ -1,13 +1,19 @@
 package com.example.demo.services;
 
+import com.example.demo.exceptions.DuplicateEntityException;
+import com.example.demo.exceptions.EntityNotFoundException;
 import com.example.demo.models.Like;
 import com.example.demo.models.Post;
 import com.example.demo.models.User;
 import com.example.demo.repositories.LikeRepository;
-import com.example.demo.repositories.PostRepository;
 import com.example.demo.services.interfaces.LikeService;
+import com.example.demo.services.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.security.Principal;
+import java.sql.SQLException;
+import java.sql.SQLIntegrityConstraintViolationException;
 
 @Service
 public class LikeServiceImpl implements LikeService {
@@ -21,6 +27,9 @@ public class LikeServiceImpl implements LikeService {
 
     @Override
     public void createLike(User user, Post post) {
+        if (checkLikeExists(user.getId(), post.getId())) {
+            throw new DuplicateEntityException("");
+        }
         Like like = new Like();
         like.setUser(user);
         like.setPost(post);
@@ -28,21 +37,26 @@ public class LikeServiceImpl implements LikeService {
         likeRepository.save(like);
     }
 
+
     @Override
-    public void deleteLike(Like like, Post post) {
+    public void deleteLike(Like like, Post post, User user) {
+        if (!checkLikeExists(user.getId(), post.getId())) {
+            throw new EntityNotFoundException("");
+        }
         post.removeLike(like);
-        post.setLiked(false);
         likeRepository.delete(like);
     }
 
     @Override
-    public Like getLikeByUserIdAndPostId(long userId, long postId) {
-        return likeRepository.getLikeByUser_IdAndPost_Id(userId, postId);
+    public boolean checkLikeExists(long userId, long potsId) {
+        Like like = likeRepository.getLikeByUser_IdAndPost_Id(userId, potsId);
+        return like != null;
     }
 
+
     @Override
-    public boolean checkIfLikeExist(long userId, long postId) {
-        return likeRepository.existsByUniquePair(userId, postId);
+    public Like getLikeByUserIdAndPostId(long userId, long postId) {
+        return likeRepository.getLikeByUser_IdAndPost_Id(userId, postId);
     }
 
     @Override
